# fs: add c++ fast path for writeFileSync string utf8