@CanadaHonk 2023-10-11
Sorry for inactivity been busy, updated for all comments said and should be good now :)