@anonrig 2023-10-11
@santigimeno Can you review again? It seems your block has been addressed.