@CanadaHonk 2023-11-19
Updated PR description after rebenchmarking, still looks like a >1.5x speedup for most situations (using Bun's benchmark).