@CanadaHonk 2023-11-21
Fixed Windows CI failing due to not giving the path to open's debug info. Should be ready now!