@RafaelGSS 2023-11-28
@CanadaHonk Do you have benchmark results from our CI? I don't we should mention local benchmarks in the changelog.