@ChALkeR 2025-11-02
base64 is also a thing, as it covers binary writes when content is already in base64 
and could also be optimized via a fast path (e.g. by adding an encoding parameter here)

```
7113870	selenium-webdriver-4.38.0.tgz/lib/webdriver.js:1742:    fs.writeFileSync(zipFilePath, Buffer.from(base64Content, 'base64'))
1563070	cucumber-html-reporter-7.2.0.tgz/lib/reporter.js:320:                  fs.writeFileSync(filename, embedding.data, 'base64');
1563070	cucumber-html-reporter-7.2.0.tgz/lib/reporter.js:334:                  fs.writeFileSync(filename, embedding.data, 'base64');
313200	@loki/diff-looks-same-0.35.0.tgz/src/create-looks-same-differ.spec.js:13:  fs.writeFileSync(outputPath, imageData, { encoding: 'base64' });
307699	@loki/diff-pixelmatch-0.35.0.tgz/src/create-pixelmatch-differ.spec.js:13:  fs.writeFileSync(outputPath, imageData, { encoding: 'base64' });
204239	backstopjs-6.3.25.tgz/core/util/compare/store-failed-diff-stub.js:10:  fs.writeFileSync(getFailedDiffFilename(testPath), BASE64_PNG_STUB, 'base64');
195127	vscode-extension-tester-8.19.0.tgz/out/browser.js:233:        fs.writeFileSync(path.join(dir, `${name}.png`), data, 'base64');
177664	azure-pipelines-task-lib-5.2.1.tgz/internal.js:927:        fs.writeFileSync(storageFile, encryptKey.toString('base64'), { encoding: 'utf8' });
82614	base64-img-1.0.4.tgz/base64-img.js:113:  fs.writeFileSync(filepath, result.base64, { encoding: 'base64' });
82614	base64-img-1.0.4.tgz/test/index.js:76:    fs.writeFileSync(getpath('dest/svg.html'), `<img src="${data}">`, { encoding: 'utf8'})
74527	remove.bg-1.3.0.tgz/index.js:77:                fs.writeFileSync(options.outputFile, result.body.data.result_b64, { encoding: "base64" });
73121	pxt-core-12.2.3.tgz/built/cli.js:2936:            nodeutil.writeFileSync(p, contents, { encoding: "base64" });
73121	pxt-core-12.2.3.tgz/built/pxt.js:165454:            nodeutil.writeFileSync(p, contents, { encoding: "base64" });
...
```

Left column is downloads/month